requirements.txt
attr==0.3.2
caffe==0.1.0
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==44.0.0
docutils==0.21.2
facenet_pytorch==2.6.0
fastapi==0.115.6
ffmpeg==1.4
HTMLParser==0.0.2
importlib_metadata==8.5.0
importlib_metadata==8.0.0
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.4
jnius==1.1.0
keyring==25.5.0
lockfile==0.12.2
matplotlib==3.10.0
mediapipe==0.10.18
mock==5.1.0
numpy==1.24.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.0.0
protobuf==5.29.1
pydantic==2.10.3
pydub==0.25.1
pyOpenSSL==24.3.0
railroad==0.5.0
soundfile==0.12.1
Sphinx==8.1.3
SQLAlchemy==2.0.35
thread==2.0.5
torch==2.2.2
torch==2.3.0
torchaudio==2.2.2
torchaudio==2.3.0
tornado==6.4.1
tqdm==4.65.2
truststore==0.10.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
